Attempt 2
I'm having a hard time following the github action documentation.  I suspect that it assumes a lot of knowledge about how this is all supposed to work.  Having not really done this before, I feel rather lost.  For example, the first error I got after simply copying the yml from the action I wanted and putting it in a yml file, the error I got when it ran was "No event triggers defined in `on`".  There was no `on` clause in the yml from the action I wanted, nor do any of the action documentation items say whether the action yml I get from a distributed action is to go in the same file with the `on` clause I found or where in the hierarchy I should put that yml snippet...  So I'm just trying things.  I wish there was a step by step example for starting from any specific action because I don't want other superfluous actions.  Ugh.