Attempt 4 at github action
I still got a generic yaml syntax error on line 19, so I started to replace the test script call with a simple `echo santity check`, but then I noticed that the interface underlines syntax errors with a red squiggle and when you hover over it, you get an error in a tooltip that said "cannot read an implicit mapping pair; a colon is missed".  Googling that was not directly helpful, but I found a yaml parser online and it was a bit more helpful.  Apparently, I needed to indent the commands 2 more spaces.  So here goes attempt 4...