github actions attempt 6
I got the error "No runner matching the specified labels was found: ubuntu-latest, macos-latest".  I could not find a resolution to this in the documentation or via google, but reading between the lines, I'm guessing that maybe specifying multiple OS's only works for self-hosted actions, so I'm going to try just testing with ubuntu.